+++ packages/plugin-experiment-react-native/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-beta.4](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-experiment-react-native@1.0.0-beta.3...@amplitude/plugin-experiment-react-native@1.0.0-beta.4) (2026-09-23)
+
+**Note:** Version bump only for package @amplitude/plugin-experiment-react-native
+
+
+
+
+
+# [1.0.0-beta.3](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-experiment-react-native@1.0.0-beta.2...@amplitude/plugin-experiment-react-native@1.0.0-beta.3) (2026-09-21)
+
+**Note:** Version bump only for package @amplitude/plugin-experiment-react-native
+
+
+
+
+
 # [1.0.0-beta.2](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-experiment-react-native@1.0.0-beta.1...@amplitude/plugin-experiment-react-native@1.0.0-beta.2) (2026-09-16)
 
 **Note:** Version bump only for package @amplitude/plugin-experiment-react-native