+++ packages/session-replay-browser/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.50.3](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/session-replay-browser@1.50.2...@amplitude/session-replay-browser@1.50.3) (2026-09-23)
+
+**Note:** Version bump only for package @amplitude/session-replay-browser
+
+
+
+
+
+## [1.50.2](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/session-replay-browser@1.50.1...@amplitude/session-replay-browser@1.50.2) (2026-09-21)
+
+**Note:** Version bump only for package @amplitude/session-replay-browser
+
+
+
+
+
 ## [1.50.1](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/session-replay-browser@1.50.0...@amplitude/session-replay-browser@1.50.1) (2026-09-16)
 
 **Note:** Version bump only for package @amplitude/session-replay-browser